refactor: dissolve ForkClone into Fork; Fix delegates to Fork
Conceptually the bootstrap operation IS fork — fork+clone+upstream is
what the verb does, full stop. Having a separate ForkClone class made
it look like a sibling of Fork when it's actually the same operation,
just without the CLI wrapper. Fix is fork + (issue branch + announce).

Now:

- Fork has two faces. `Fork#run(argv)` is the CLI entry. `Fork#call`
  is the public primitive (idempotent fork+clone+upstream-remote,
  returns local path). Same operation, different invocation.
- Fix takes a `fork:` kwarg (a Fork instance) and delegates to
  `@fork.call(...)` for the bootstrap step, then layers on branch
  creation and the working-on-this comment.
- ForkClone class and its spec deleted. Primitive tests merged into
  fork_spec.rb under "#call (the primitive)".

Net: 4 abstractions → 3 (Workflow, Fork, Fix). Fix's relationship to
Fork is now explicit in the code: it composes Fork.
