+++ lib/gem_contribute/cli/fork.rb
@@ -1,18 +1,23 @@
 # frozen_string_literal: true
 
+require "fileutils"
+
 module GemContribute
   module CLI
-    # `gem-contribute fork <gem> [-e] [-a]`
+    # `gem-contribute fork <gem|owner/repo> [-e] [-a]` and the underlying
+    # bootstrap primitive used by `fix`.
     #
-    # The look-around-first path: fork + clone + upstream remote, leave
-    # the user on the default branch. Use this when you want to explore
-    # a project before committing to a specific issue.
-    # `gem-contribute fix <gem>/<issue>` is the issue-tied counterpart;
-    # both compose the same `ForkClone` primitive.
+    # Two faces, same operation:
+    # - `Fork#run(argv)` — the CLI verb. Resolve, fork+clone, summary, hooks.
+    # - `Fork#call(adapter, project, viewer)` — the primitive. Idempotent
+    #   fork (or reuse) + clone (or reuse) + upstream remote, returns the
+    #   local clone path. `Fix` delegates to this for its bootstrap step.
     class Fork
       include Workflow
 
       DEFAULT_CLONE_ROOT = File.expand_path("~/code/oss")
+      FORK_READINESS_RETRIES = 12 # 12 × 5s = 60s ceiling
+      FORK_READINESS_INTERVAL = 5
 
       # rubocop:disable Metrics/ParameterLists
       def initialize(stdout: $stdout, stderr: $stderr,
@@ -21,19 +26,31 @@ module GemContribute
                      git: Git.new,
                      clone_root: DEFAULT_CLONE_ROOT,
                      sleeper: ->(s) { Kernel.sleep(s) },
-                     fork_clone: nil,
                      post_clone_hooks: nil)
         @stdout = stdout
         @stderr = stderr
         @resolver = resolver
         @store = store
         @adapter_factory = adapter_factory
+        @git = git
         @clone_root = clone_root
-        @fork_clone = fork_clone || ForkClone.new(stdout: stdout, git: git,
-                                                  clone_root: clone_root, sleeper: sleeper)
+        @sleeper = sleeper
         @post_clone_hooks = post_clone_hooks || PostCloneHooks.new(stdout: stdout, stderr: stderr)
       end
       # rubocop:enable Metrics/ParameterLists
+
+      # Public primitive. Idempotent: reuses an existing fork (via
+      # `already_forked?`) and an existing local clone (via the `.git`
+      # directory). Returns the local clone path.
+      def call(adapter, project, viewer)
+        clone_url = ensure_fork(adapter, project, viewer)
+        local_path = clone_into_root(project, clone_url)
+        # `submit` needs to know the canonical project to point the PR at.
+        # `upstream` follows the standard fork workflow convention.
+        @git.add_remote(local_path, "upstream",
+                        "https://github.com/#{project.owner}/#{project.repo}.git")
+        local_path
+      end
 
       def run(argv)
         with_workflow_rescues("fork") do
@@ -74,7 +91,7 @@ module GemContribute
 
       def execute(adapter, project, flags)
         viewer = adapter.viewer_login
-        local_path = @fork_clone.call(adapter, project, viewer)
+        local_path = call(adapter, project, viewer)
 
         print_summary(local_path, project, viewer)
         @post_clone_hooks.call(local_path, editor: flags[:editor], ai_tool: flags[:ai_tool])
@@ -90,6 +107,46 @@ module GemContribute
         @stdout.puts "Next: cd #{local_path} && explore. When you pick an issue, " \
                      "`gem-contribute fix #{project.gem_name}/<issue#>` " \
                      "branches off the default."
+      end
+
+      # === Primitive helpers (private) ===
+
+      def ensure_fork(adapter, project, viewer)
+        if adapter.already_forked?(project)
+          @stdout.puts "You already have a fork at #{viewer}/#{project.repo}. Skipping fork."
+          return "https://github.com/#{viewer}/#{project.repo}.git"
+        end
+
+        @stdout.puts "Forking #{project.owner}/#{project.repo} → #{viewer}/#{project.repo}..."
+        body = adapter.fork(project)
+        wait_until_ready(adapter, viewer, project.repo)
+        body.fetch("clone_url")
+      end
+
+      def wait_until_ready(adapter, viewer, name)
+        ready = FORK_READINESS_RETRIES.times.any? do |i|
+          break true if adapter.fork_ready?(viewer, name)
+
+          @sleeper.call(FORK_READINESS_INTERVAL) unless i == FORK_READINESS_RETRIES - 1
+          false
+        end
+        return if ready
+
+        raise GemContribute::AdapterError,
+              "fork not reachable after #{FORK_READINESS_RETRIES * FORK_READINESS_INTERVAL}s"
+      end
+
+      def clone_into_root(project, clone_url)
+        target = File.join(@clone_root, project.owner, project.repo)
+        if File.directory?(File.join(target, ".git"))
+          @stdout.puts "Reusing existing clone at #{target}."
+          return target
+        end
+
+        FileUtils.mkdir_p(File.dirname(target))
+        @stdout.puts "Cloning into #{target}..."
+        @git.clone(clone_url, target)
+        target
       end
     end
   end
